Rename site-wide sut commands
